Fix writeNSS passwd, shadow, and group lines
Between 6.2.10 and 7.0.0, changes to `writeNSS` prevented the
`users` variable being set to anything other than "", so the
`passwd`, `shadow` and `group` lines written by authconfig would
always have no values.

This fixes that, and rejigs the code a little to be clearer and
avoid needless initialization of variables to the empty string
and leading spaces.