Fix incorrect PAM configuration with SSSD (CVE-2017-7488).
The PAM configuration that authconfig previously generated since
the release 7.0.0 when pam_sss was enabled caused information
about the presence or non-presence of user account on the system
to be leaked via different prompting and/or timing.

The 'authconfig --updateall' must be run to enforce regeneration
of the PAM configuration files to fix the incorrect configuration.